Add gulp-plumber to prevent pipe from breaking on compile error
Currently, when running the `gulp watch` task, a syntax error in
one of the source files causes the watch task to break and the
task has to be cancelled and restarted.

With this update, an error will just terminate the current pipe
but allow the watch task to run properly the next time there are
file changes.